backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
pandas
openpyxl
python-multipart
pdf2image
pillow
pydantic
python-dotenv
